// frontend/src/shared/api/generated/errors.ts
// Code generated by tygo. DO NOT EDIT.

//////////
// source: errors.go

export type ErrorCode = string;
export const EntityNotFoundError: ErrorCode = "entityNotFoundError";
export const EntityCreateError: ErrorCode = "entityCreateError";
export const EntityUpdateError: ErrorCode = "entityUpdateError";
export const EntityDeleteError: ErrorCode = "entityDeleteError";
export const ValidationError: ErrorCode = "validationError";
export const UnauthorizedError: ErrorCode = "unauthorizedError";
export const ForbiddenError: ErrorCode = "forbiddenError";
export const InternalError: ErrorCode = "internalError";
export interface ValidationDetail {
  instancePath: string;
  message: string;
  keyword: string;
}
export interface AppError {
  code: ErrorCode;
  statusCode: number /* int */;
  message: string;
  details?: ValidationDetail[];
}